ml_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pymongo==4.5.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
google-generativeai==0.3.0
shap==0.44.0
lightgbm==4.1.0
asyncio==3.4.3
httpx==0.25.1
yfinance==0.2.36
feedparser==6.0.10
ta==0.11.0
